+++ src/finchlite/autoschedule/__init__.py
@@ -1,9 +1,18 @@
+from finchlite.autoschedule.optimize import DefaultLogicOptimizer
+from finchlite.finch_notation.interpreter import NotationInterpreter
+
+from ..codegen import NumbaCompiler
+from ..compile import NotationCompiler
+from ..finch_assembly import AssemblyInterpreter, AssemblySimplify
+from .galley_optimize import GalleyLogicalOptimizer
 from ..finch_logic import (
     Aggregate,
     Alias,
     Field,
     Literal,
+    LogicInterpreter,
     MapJoin,
+    MockLogicLoader,
     Plan,
     Produces,
     Query,
@@ -17,11 +26,70 @@ from .compiler import LogicCompiler, NotationGenerator
 from .executor import LogicExecutor
 from .formatter import DefaultLogicFormatter, LogicFormatter
 from .normalize import LogicNormalizer, normalize_names
-from .optimize import DefaultLogicOptimizer
 from .stages import LogicEinsumLowerer, LogicNotationLowerer
 from .standardize import LogicStandardizer
 
+INTERPRET_LOGIC = LogicInterpreter()
+OPTIMIZE_LOGIC = LogicNormalizer(
+    LogicExecutor(
+        DefaultLogicOptimizer(
+            LogicStandardizer(DefaultLogicFormatter(MockLogicLoader()))
+        )
+    )
+)
+INTERPRET_NOTATION = LogicNormalizer(
+    LogicExecutor(
+        DefaultLogicOptimizer(
+            LogicStandardizer(
+                DefaultLogicFormatter(LogicCompiler(NotationInterpreter()))
+            )
+        )
+    )
+)
+INTERPRET_ASSEMBLY = LogicNormalizer(
+    LogicExecutor(
+        DefaultLogicOptimizer(
+            LogicStandardizer(
+                DefaultLogicFormatter(
+                    LogicCompiler(NotationCompiler(AssemblyInterpreter()))
+                )
+            )
+        )
+    )
+)
+COMPILE_NUMBA = LogicNormalizer(
+    LogicExecutor(
+        DefaultLogicOptimizer(
+            LogicStandardizer(
+                DefaultLogicFormatter(
+                    LogicCompiler(
+                        NotationCompiler(
+                            NumbaCompiler(), ctx_transforms=(AssemblySimplify(),)
+                        )
+                    )
+                )
+            )
+        )
+    )
+)
+
+INTERPRET_NOTATION_GALLEY = LogicNormalizer(
+    LogicExecutor(
+        GalleyLogicalOptimizer(
+            LogicStandardizer(
+                DefaultLogicFormatter(LogicCompiler(NotationInterpreter()))
+            )
+        )
+    )
+)
+
 __all__ = [
+    "COMPILE_NUMBA",
+    "INTERPRET_ASSEMBLY",
+    "INTERPRET_LOGIC",
+    "INTERPRET_NOTATION",
+    "INTERPRET_NOTATION_GALLEY",
+    "OPTIMIZE_LOGIC",
     "Aggregate",
     "Alias",
     "DefaultLogicFormatter",
