+++ src/finchlite/interface/__init__.py
@@ -99,12 +99,6 @@ from .eager import (
     vecdot,
 )
 from .fuse import (
-    COMPILE_NUMBA,
-    INTERPRET_ASSEMBLY,
-    INTERPRET_LOGIC,
-    INTERPRET_NOTATION,
-    INTERPRET_NOTATION_GALLEY,
-    OPTIMIZE_LOGIC,
     compute,
     fuse,
     fused,
@@ -116,12 +110,6 @@ from .np_wrapper import NumPyFType, NumPyWrapper
 from .scalar import Scalar, ScalarFType
 
 __all__ = [
-    "COMPILE_NUMBA",
-    "INTERPRET_ASSEMBLY",
-    "INTERPRET_LOGIC",
-    "INTERPRET_NOTATION",
-    "INTERPRET_NOTATION_GALLEY",
-    "OPTIMIZE_LOGIC",
     "EagerTensor",
     "LazyTensor",
     "NumPyFType",
