+++ tests/conftest.py
@@ -7,8 +7,8 @@ from numpy.random import default_rng
 from numpy.testing import assert_allclose, assert_equal
 
 from finchlite import get_default_scheduler, set_default_scheduler
+from finchlite.autoschedule import COMPILE_NUMBA, INTERPRET_LOGIC
 from finchlite.finch_logic import Field
-from finchlite.interface.fuse import COMPILE_NUMBA, INTERPRET_LOGIC
 
 
 @pytest.fixture(scope="session")
